// lib/api/generated/models/loginResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 0.1.0
 */
import type { PublicUser } from './publicUser';

export interface LoginResponse {
  expires_in_hours: number;
  token: string;
  token_type: string;
  user: PublicUser;
}