// lib/api/generated/models/role.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 0.1.0
 */

/**
 * Role klubowe — można łączyć (np. zawodnik + trener).
 */
export type Role = typeof Role[keyof typeof Role];


export const Role = {
  zawodnik: 'zawodnik',
  trener: 'trener',
  admin: 'admin',
  superadmin: 'superadmin',
} as const;